# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Cube-back",
        "version": "1.0.0"
    },
    "paths": {
        "/api/categories": {
            "get": {
                "tags": [
                    "Categorie"
                ],
                "summary": "Récupérer toutes les ressources",
                "operationId": "d753b5a85d077f3a311df9eb28de8a1a",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items to return per page (1-100)",
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Keyword for searching items",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort items by",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order (asc or desc)",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorie"
                ],
                "summary": "Créer une catégorie",
                "operationId": "1d6b893ef899e2aecece0a5e18049d8a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "intitule_cat"
                                ],
                                "properties": {
                                    "intitule_cat": {
                                        "type": "string",
                                        "maxLength": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categorie"
                ],
                "summary": "Récupérer une catégorie spécifique",
                "operationId": "02f8d50a129f2d0b43f583dfe51c9cc5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the item",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Item not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categorie"
                ],
                "summary": "Modifier une catégorie",
                "operationId": "2eaa317f2ea8871f48bcecdfe95e6a83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the item",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "intitule_cat"
                                ],
                                "properties": {
                                    "intitule_cat": {
                                        "type": "string",
                                        "maxLength": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item updated successfully"
                    },
                    "404": {
                        "description": "Item not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categorie"
                ],
                "summary": "Supprimer une catégorie",
                "operationId": "21702b2a8b7abdf7a40668fdcdf63a16",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the item",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item deleted successfully"
                    },
                    "404": {
                        "description": "Item not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}